microservices/ResumeProcessing/requirements.txt
pika==1.2.1
python-docx
PyPDF2
ollama
